move classes to /src
